Extract samples into base OsuSliderBar.cs from NormalSliderBar.cs
